Initial commit
Created library implementing dictionary functionality in C. The term
"dictionary" was chosen in reminiscence of Python's "dict" data type, due to the
fact that a Python dict can contain key-value pairs of many different types, a
feature this library shares. (A feature, I may add, which arose kind of
accidentally, but meh.)

I realize that there are probably many (a lot better) implementations of
hashtables and such things out on the interwebz, but that's beside the point. I
started this project to get a better understanding of the inner workings of
hashtables, and I feel like I've accomplished that. However, I'll probably at
least use this library somewhat myself, because _hashtables are friggin'
useful_. So there.

I'd like to further improve this library on some points. First, I'd like to
create an actual lib file which I can link against. However, I probably also
need to take a serious look at efficiency and/or memory management at some
point, because I'm not confident that I've nailed those at all. I'd also like to
be able to implement a self-balancing tree bucket structure as an (optional?)
alternative to the linked list currently in place. But I'm not sure how
difficult that will be. Also, I'll probably have to fix some bugs.

That's it for now. It's in your hands now, future me. Good luck!