Replace makefile with automake scripts
For simplicity on my own part, I've just added the files which generate the
configure and Makefile.in files to the git repo. However, it's probably better
to add the generated configure and Makefile.in files instead, since they (as far
as I've gathered) don't require autotools to be on the computer of the person
who builds the library. I need to work out how (or even whether) to install the
library with make install before I feel confident in letting those configuration
files go.

For now, I've also added a build.sh file which just runs all the steps required
to generate and build everything. I should also add some script for cleanup,
because as it is there are a lot of generated files which clutter everything.