Remove SUBDIRS line
Looks like SUBDIRS isn't actually necessary to make it work. I just came across
something similar earlier and included it, then later on I got it to work so I
just stuck with the configuration I had back then.

NOTE: For the record, I've added RANLIB in a different way than "normal",
because I had issues with it not being defined. If I do it like this, it looks
like it works, so I'm sticking with it.