Decentralize makefile operations, add example
I've moved the responsibility of making the files to the end directories in
which they're supposed to end up. In other words, the logic for making the
library is in a Makefile.am file in the lib folder, etc. Due to this I also had
to update .gitignore to not ignore the lib folder entirely, because it should
not ignore the Makefile.am file there.

I also added an example of how to use the CDict library. It simply reads a lot
of lines from stdin, and keeps a count in the dictionary for how many times each
line has been read. In the end it prints out the count for each string it
recorded. I'll probably add more examples later.

The way I had to loop through the files makes me think it'd be useful to keep a
list of keys, because that means we only need to go through that list, meaning
that it should be roughly O(n) complexity, where n is the number of distinct
keys, since lookup should be around O(1) on average. I also kind of want to
change the API so that dict_new takes no arguments and returns a dictionary with
reasonable default values, with a separate function for creating a dictionary
with a given size and load factor, because most of the time I just want stuff to
_work_, dammit!