Add array of keys to simplify loopthrough
The user technically has both the option of looping through this array and to go
through each bucket and traverse the linked list if necessary. If buckets are
deep, the second option might be a bit faster, since it won't have to traverse
the same linked list more than once. However, normally it'll probably be a
negligible difference, and the former looks a lot more intuitive.