Add dict_get_t macro
Added a macro for more easily accessing the element in the dictionary. The
dict_get_t macro takes a dictionary, a type, and a key, and returns the
corresponding value cast to the correct type. The drawback is that this macro is
a HORRIBLE HACK which should NEVER BE USED and MAY STOP WORKING AT ANY TIME
WITHOUT WARNING. Ye have been warned.